test: split apply promotion tests
