+++ test/clawsweeper.test.ts
@@ -85,10 +85,10 @@ import {
   implementedCloseReport,
   item,
   lowSignalCloseReport,
-  markedReviewCommentForTest,
   promotionGhMock,
   reportWithSyncedReviewComment,
   runApplyDecisionsForTest,
+  stalePullRequestReport,
   tmpPrefix,
   withMockCodexProof,
   withMockGh,
@@ -1238,2246 +1238,6 @@ if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/321\\/timeline(?:\\?|$
 }
 `;
 }
-
-function stalePullRequestReport(overrides = {}) {
-  return `${workPlanCandidateReport({
-    repository: "openclaw/openclaw",
-    number: 330,
-    type: "pull_request",
-    title: "Stale F PR",
-    url: "https://github.com/openclaw/openclaw/pull/330",
-    author: "reporter",
-    author_association: "CONTRIBUTOR",
-    decision: "keep_open",
-    close_reason: "none",
-    confidence: "high",
-    action_taken: "kept_open",
-    work_candidate: "manual_review",
-    work_status: "manual_review",
-    item_snapshot_hash: "reviewed-snapshot",
-    item_created_at: "2026-02-01T00:00:00Z",
-    item_updated_at: "2026-05-01T00:00:00Z",
-    reviewed_at: "2026-05-01T00:00:00Z",
-    labels: JSON.stringify(["status: 📣 needs proof"]),
-    pr_rating_overall: "F",
-    pr_rating_proof: "F",
-    pr_rating_patch: "F",
-    ...overrides,
-  })}\n\n## Real Behavior Proof\n\nStatus: missing\nEvidence kind: none\nNeeds contributor action: true\nSummary: No live proof was supplied.\n\n## PR Rating\n\nOverall tier: F\nProof tier: F\nPatch tier: F\nSummary: The PR is not merge-ready.\nNext rank-up steps:\n- Rebase and provide proof.\n`;
-}
-
-function stripProofAndRatingFrontMatter(report: string): string {
-  return report.replace(
-    /\n(?:real_behavior_proof_status|pr_rating_overall|pr_rating_proof|pr_rating_patch):[^\n]*/g,
-    "",
-  );
-}
-
-test("apply-decisions upgrades live no-diff kept-open PRs to duplicate closes", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    writeFileSync(
-      join(itemsDir, "322.md"),
-      workPlanCandidateReport({
-        number: 322,
-        repository: "openclaw/openclaw",
-        type: "pull_request",
-        title: "Empty PR",
-        url: "https://github.com/openclaw/openclaw/pull/322",
-        decision: "keep_open",
-        close_reason: "none",
-        action_taken: "kept_open",
-        item_snapshot_hash: "reviewed-snapshot",
-        item_created_at: "2026-05-01T00:00:00Z",
-        item_updated_at: "2026-05-01T00:00:00Z",
-        pull_head_sha: "head-sha",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }),
-      "utf8",
-    );
-
-    const ghMock = `
-const rawArgs = process.argv.slice(2);
-const args = rawArgs[0] === "--repo" ? rawArgs.slice(2) : rawArgs;
-const path = args[1] || "";
-if (args[0] === "api" && args[1] === "-i" && /\\/issues\\/322\\/timeline(?:\\?|$)/.test(args[2] || "")) {
-  console.log("HTTP/2 200\\n\\n[]");
-} else if (args[0] === "api" && /\\/issues\\/322\\/comments(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/322\\/timeline(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "api" && /\\/issues\\/322$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 322,
-    title: "Empty PR",
-    html_url: "https://github.com/openclaw/openclaw/pull/322",
-    body: "No remaining diff.",
-    created_at: "2026-05-01T00:00:00Z",
-    updated_at: "2026-05-01T00:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "reporter" },
-    labels: [],
-    comments: 0,
-    pull_request: { url: "https://api.github.com/repos/openclaw/openclaw/pulls/322" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/322$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 322,
-    title: "Empty PR",
-    html_url: "https://github.com/openclaw/openclaw/pull/322",
-    state: "open",
-    changed_files: 0,
-    commits: 0,
-    review_comments: 0,
-    body: "No remaining diff.",
-    head: { sha: "head-sha", ref: "branch", repo: { full_name: "fork/openclaw" } },
-    base: { sha: "base-sha", ref: "main", repo: { full_name: "openclaw/openclaw" } },
-    user: { login: "reporter" }
-  }));
-} else if (args[0] === "api" && /\\/pulls\\/400$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 400,
-    title: "Old related PR",
-    html_url: "https://github.com/openclaw/openclaw/pull/400",
-    state: "closed",
-    merged_at: "2026-05-02T00:00:00Z",
-    updated_at: "2026-05-02T00:00:00Z",
-    mergeable_state: "clean",
-    body: "Old related PR body.",
-    labels: [{ name: "proof: sufficient" }]
-  }));
-} else if (args[0] === "api" && /\\/issues\\/400$/.test(path)) {
-  console.log(JSON.stringify({
-    number: 400,
-    title: "Old related PR",
-    html_url: "https://github.com/openclaw/openclaw/pull/400",
-    body: "Old related PR body.",
-    state: "closed",
-    updated_at: "2026-05-02T00:00:00Z",
-    labels: [{ name: "proof: sufficient" }],
-    comments: 0,
-    pull_request: { url: "https://api.github.com/repos/openclaw/openclaw/pulls/400" }
-  }));
-} else if (args[0] === "api" && /\\/issues\\/400\\/comments(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([]));
-} else if (args[0] === "api" && /\\/pulls\\/400\\/files(?:\\?|$)/.test(path)) {
-  const files = [{ filename: "src/runtime.ts" }];
-  if (args.includes("--jq")) console.log(JSON.stringify(files.map((file) => file.filename)));
-  else console.log(JSON.stringify([files]));
-} else if (args[0] === "api" && /\\/pulls\\/322\\/(files|commits|comments)(?:\\?|$)/.test(path)) {
-  console.log(JSON.stringify([[]]));
-} else if (args[0] === "label" || args[0] === "issue") {
-  console.log("");
-} else {
-  console.error("unexpected gh args", JSON.stringify(args));
-  process.exit(1);
-}
-`;
-    withMockGh(root, ghMock, () => {
-      withMockCodexProof(
-        root,
-        { type: "failure", message: "proof should not run for no-diff PR" },
-        () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        },
-      );
-    });
-
-    assert.deepEqual(JSON.parse(readFileSync(reportPath, "utf8")), [
-      {
-        number: 322,
-        action: "review_comment_synced",
-        reason: "would create durable Codex review comment",
-      },
-      {
-        number: 322,
-        action: "closed",
-        reason:
-          "dry-run: would close as duplicate or superseded; dry-run: would post close-applied comment",
-      },
-    ]);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions promotes old F-rated stale PRs to duplicate closes", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      stalePullRequestReport({
-        work_cluster_refs: JSON.stringify(["Related discussion in #400"]),
-      }),
-      330,
-      "none",
-    );
-    writeFileSync(join(itemsDir, "330.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 330,
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Related cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Related cleanup, not stale PR coverage evidence.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          { type: "failure", message: "proof should not run for stale promotion incidental ref" },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      number: number;
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "review_comment_synced"),
-      true,
-    );
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "closed")?.reason ?? "",
-      /duplicate or superseded/,
-    );
-    assert.doesNotMatch(JSON.stringify(report), /proof should not run/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions promotes stale PRs after automation-only drift", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(stalePullRequestReport(), 330, "none");
-    writeFileSync(join(itemsDir, "330.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 330,
-        comment: synced.comment,
-        itemUpdatedAt: "2026-05-02T00:00:00Z",
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B is the canonical PR covering PR A.",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote stale PRs from truncated activity", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(stalePullRequestReport(), 330, "none");
-    writeFileSync(join(itemsDir, "330.md"), synced.report, "utf8");
-
-    const comments = Array.from({ length: 24 }, (_, index) => ({
-      id: 9330 + index,
-      html_url: `https://github.com/openclaw/openclaw/pull/330#issuecomment-${9330 + index}`,
-      created_at: "2026-05-01T01:00:00Z",
-      updated_at: "2026-05-01T01:00:00Z",
-      user: { login: "clawsweeper[bot]" },
-      body: index === 0 ? synced.comment : "automation label sync",
-    }));
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 330,
-        comment: synced.comment,
-        comments,
-        issueCommentCount: 25,
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "proof should not run" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.doesNotMatch(JSON.stringify(report), /proof should not run/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote stale PRs after human follow-up", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(stalePullRequestReport(), 330, "none");
-    writeFileSync(join(itemsDir, "330.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 330,
-        comment: synced.comment,
-        comments: [
-          {
-            id: 9330,
-            html_url: "https://github.com/openclaw/openclaw/pull/330#issuecomment-9330",
-            created_at: "2026-05-01T01:00:00Z",
-            updated_at: "2026-05-01T01:00:00Z",
-            user: { login: "clawsweeper[bot]" },
-            body: synced.comment,
-          },
-          {
-            id: 9331,
-            html_url: "https://github.com/openclaw/openclaw/pull/330#issuecomment-9331",
-            created_at: "2026-05-01T02:00:00Z",
-            updated_at: "2026-05-01T02:00:00Z",
-            user: { login: "reporter" },
-            body: "I can still work on this.",
-          },
-        ],
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "proof should not run" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.doesNotMatch(JSON.stringify(report), /proof should not run/);
-    assert.match(readFileSync(join(itemsDir, "330.md"), "utf8"), /^action_taken: kept_open$/m);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions promotes recommended pause-or-close PRs", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      stalePullRequestReport({
-        number: 331,
-        title: "Superseded prompt PR",
-        pr_rating_overall: "D",
-        pr_rating_proof: "D",
-        merge_risk_options: JSON.stringify([
-          {
-            title: "Close as superseded after maintainer decision",
-            body: "Current-main prompt work already covers the useful guidance.",
-            category: "pause_or_close",
-            recommended: true,
-            automergeInstruction: "",
-          },
-        ]),
-      }),
-      331,
-      "none",
-    );
-    writeFileSync(join(itemsDir, "331.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({ number: 331, title: "Superseded prompt PR", comment: synced.comment }),
-      () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: [
-            "--target-repo",
-            "openclaw/openclaw",
-            "--dry-run",
-            "--apply-kind",
-            "all",
-            "--processed-limit",
-            "3",
-          ],
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions promotes PRs superseded by linked pull requests", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const linkedMarkdownLabelReport = stalePullRequestReport({
-      number: 332,
-      title: "Old activity PR",
-      pr_rating_overall: "D",
-      pr_rating_proof: "D",
-      pr_rating_patch: "D",
-      work_cluster_refs: JSON.stringify([
-        "[replacement PR](https://github.com/openclaw/openclaw/pull/400)",
-      ]),
-    })
-      .replace("Overall tier: F", "Overall tier: D")
-      .replace("Proof tier: F", "Proof tier: D")
-      .replace("Patch tier: F", "Patch tier: D");
-    const synced = reportWithSyncedReviewComment(linkedMarkdownLabelReport, 332, "none");
-    writeFileSync(join(itemsDir, "332.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 332,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Canonical activity PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-            labels: ["proof: sufficient"],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B is the canonical PR covering PR A.",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      number: number;
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "closed")?.reason ?? "",
-      /duplicate or superseded/,
-    );
-    assert.match(readFileSync(proofLogPath, "utf8"), /proof/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote docs-only PRs superseded by code-only pull requests", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      stalePullRequestReport({
-        number: 337,
-        title: "ENETDOWN docs companion",
-        pr_rating_overall: "D",
-        pr_rating_proof: "D",
-        pr_rating_patch: "D",
-        pull_files: JSON.stringify(["docs/gateway/troubleshooting.md", "docs/platforms/macos.md"]),
-        pull_files_truncated: false,
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }),
-      337,
-      "none",
-    );
-    writeFileSync(join(itemsDir, "337.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 337,
-        title: "ENETDOWN docs companion",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Canonical ENETDOWN runtime fix",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-26T17:40:32Z",
-            mergeable_state: "clean",
-            labels: ["proof: sufficient"],
-            files: [
-              "src/infra/unhandled-rejections.ts",
-              "extensions/telegram/src/network-errors.ts",
-            ],
-          },
-        },
-      }),
-      () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: [
-            "--target-repo",
-            "openclaw/openclaw",
-            "--dry-run",
-            "--apply-kind",
-            "all",
-            "--processed-limit",
-            "3",
-          ],
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions keeps promoted PR close proposals open when coverage proof fails", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      stalePullRequestReport({
-        number: 352,
-        title: "Old activity PR",
-        pr_rating_overall: "D",
-        pr_rating_proof: "D",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }),
-      352,
-      "none",
-    );
-    writeFileSync(join(itemsDir, "352.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 352,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Canonical activity PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-            labels: ["proof: sufficient"],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "model unavailable" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "retry_pr_close_coverage_proof")?.reason ?? "",
-      /PR close coverage proof failed/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions promotes PRs superseded by merged linked pull requests without proof labels", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      stalePullRequestReport({
-        number: 333,
-        title: "Old merged-replacement PR",
-        pr_rating_overall: "D",
-        pr_rating_proof: "D",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }),
-      333,
-      "none",
-    );
-    writeFileSync(join(itemsDir, "333.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 333,
-        title: "Old merged-replacement PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Merged canonical PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            mergeable_state: "dirty",
-            labels: ["status: needs proof", "rating: unranked krab"],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B is the merged canonical PR covering PR A.",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "closed")?.reason ?? "",
-      /duplicate or superseded/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote PRs superseded by no-proof linked pull requests", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      stalePullRequestReport({
-        number: 334,
-        title: "Old activity PR",
-        pr_rating_overall: "D",
-        pr_rating_proof: "D",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }),
-      334,
-      "none",
-    );
-    writeFileSync(join(itemsDir, "334.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 334,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Canonical activity PR without proof",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "proof should not run" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.doesNotMatch(JSON.stringify(report), /proof should not run/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote PRs superseded by unsafe linked pull requests", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      stalePullRequestReport({
-        number: 335,
-        title: "Old activity PR",
-        pr_rating_overall: "D",
-        pr_rating_proof: "D",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }),
-      335,
-      "none",
-    );
-    writeFileSync(join(itemsDir, "335.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 335,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Unsafe canonical PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-            labels: ["triage: needs-real-behavior-proof", "status: 📣 needs proof"],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "proof should not run" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.doesNotMatch(JSON.stringify(report), /proof should not run/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote PRs superseded by F-rated linked pull requests", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const sourceReport = stalePullRequestReport({
-      number: 338,
-      title: "Old activity PR",
-      labels: JSON.stringify([]),
-      pr_rating_overall: "D",
-      pr_rating_proof: "D",
-      pr_rating_patch: "D",
-      work_cluster_refs: JSON.stringify([
-        "Superseded by https://github.com/openclaw/openclaw/pull/400",
-      ]),
-    })
-      .replace("Status: missing", "Status: sufficient")
-      .replace(
-        "Overall tier: F\nProof tier: F\nPatch tier: F",
-        "Overall tier: D\nProof tier: D\nPatch tier: D",
-      );
-    const synced = reportWithSyncedReviewComment(sourceReport, 338, "none");
-    writeFileSync(join(itemsDir, "338.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 338,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "F-rated canonical PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-            labels: ["proof: sufficient", "rating: unranked krab"],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "proof should not run" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.doesNotMatch(JSON.stringify(report), /proof should not run/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote PRs superseded by section-only unsafe linked reports", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const sourceReport = stalePullRequestReport({
-      number: 340,
-      title: "Old activity PR",
-      labels: JSON.stringify([]),
-      pr_rating_overall: "D",
-      pr_rating_proof: "D",
-      pr_rating_patch: "D",
-      work_cluster_refs: JSON.stringify([
-        "Superseded by https://github.com/openclaw/openclaw/pull/400",
-      ]),
-    })
-      .replace("Status: missing", "Status: sufficient")
-      .replace(
-        "Overall tier: F\nProof tier: F\nPatch tier: F",
-        "Overall tier: D\nProof tier: D\nPatch tier: D",
-      );
-    const synced = reportWithSyncedReviewComment(sourceReport, 340, "none");
-    writeFileSync(join(itemsDir, "340.md"), synced.report, "utf8");
-    writeFileSync(
-      join(itemsDir, "400.md"),
-      stripProofAndRatingFrontMatter(
-        stalePullRequestReport({
-          number: 400,
-          title: "Canonical PR with old section-only blockers",
-          labels: JSON.stringify([]),
-        }),
-      ),
-      "utf8",
-    );
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 340,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Canonical PR with old section-only blockers",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "proof should not run" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--dry-run",
-              "--apply-kind",
-              "all",
-              "--item-numbers",
-              "340",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.doesNotMatch(JSON.stringify(report), /proof should not run/);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions promotes PRs when live proof labels supersede stale linked reports", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const sourceReport = stalePullRequestReport({
-      number: 342,
-      title: "Old activity PR",
-      labels: JSON.stringify([]),
-      pr_rating_overall: "D",
-      pr_rating_proof: "D",
-      pr_rating_patch: "D",
-      work_cluster_refs: JSON.stringify([
-        "Superseded by https://github.com/openclaw/openclaw/pull/400",
-      ]),
-    })
-      .replace("Status: missing", "Status: sufficient")
-      .replace(
-        "Overall tier: F\nProof tier: F\nPatch tier: F",
-        "Overall tier: D\nProof tier: D\nPatch tier: D",
-      );
-    const synced = reportWithSyncedReviewComment(sourceReport, 342, "none");
-    writeFileSync(join(itemsDir, "342.md"), synced.report, "utf8");
-    writeFileSync(
-      join(itemsDir, "400.md"),
-      stalePullRequestReport({
-        number: 400,
-        title: "Canonical PR with stale proof report",
-        labels: JSON.stringify(["status: needs proof"]),
-        pr_rating_overall: "D",
-        pr_rating_proof: "D",
-        pr_rating_patch: "D",
-      }).replace(
-        "Overall tier: F\nProof tier: F\nPatch tier: F",
-        "Overall tier: D\nProof tier: D\nPatch tier: D",
-      ),
-      "utf8",
-    );
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 342,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Canonical PR with live proof label",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-            labels: ["proof: sufficient"],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "covered",
-            reason: "PR B is the live proof-backed canonical PR covering PR A.",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--dry-run",
-                "--apply-kind",
-                "all",
-                "--item-numbers",
-                "342",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      true,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote PRs when live labels supersede stale proof reports", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const sourceReport = stalePullRequestReport({
-      number: 344,
-      title: "Old activity PR",
-      labels: JSON.stringify([]),
-      pr_rating_overall: "D",
-      pr_rating_proof: "D",
-      pr_rating_patch: "D",
-      work_cluster_refs: JSON.stringify([
-        "Superseded by https://github.com/openclaw/openclaw/pull/400",
-      ]),
-    })
-      .replace("Status: missing", "Status: sufficient")
-      .replace(
-        "Overall tier: F\nProof tier: F\nPatch tier: F",
-        "Overall tier: D\nProof tier: D\nPatch tier: D",
-      );
-    const synced = reportWithSyncedReviewComment(sourceReport, 344, "none");
-    writeFileSync(join(itemsDir, "344.md"), synced.report, "utf8");
-    writeFileSync(
-      join(itemsDir, "400.md"),
-      stalePullRequestReport({
-        number: 400,
-        title: "Canonical PR with stale sufficient proof report",
-        labels: JSON.stringify(["proof: sufficient"]),
-        pr_rating_overall: "D",
-        pr_rating_proof: "D",
-        pr_rating_patch: "D",
-      })
-        .replace("Status: missing", "Status: sufficient")
-        .replace(
-          "Overall tier: F\nProof tier: F\nPatch tier: F",
-          "Overall tier: D\nProof tier: D\nPatch tier: D",
-        ),
-      "utf8",
-    );
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 344,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Canonical PR with current needs-proof labels",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "clean",
-            labels: ["triage: needs-real-behavior-proof", "status: needs proof"],
-          },
-        },
-      }),
-      () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: [
-            "--target-repo",
-            "openclaw/openclaw",
-            "--dry-run",
-            "--apply-kind",
-            "all",
-            "--item-numbers",
-            "344",
-            "--processed-limit",
-            "3",
-          ],
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote PRs superseded by unknown-mergeability PRs", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const sourceReport = stalePullRequestReport({
-      number: 343,
-      title: "Old activity PR",
-      labels: JSON.stringify([]),
-      pr_rating_overall: "D",
-      pr_rating_proof: "D",
-      pr_rating_patch: "D",
-      work_cluster_refs: JSON.stringify([
-        "Superseded by https://github.com/openclaw/openclaw/pull/400",
-      ]),
-    })
-      .replace("Status: missing", "Status: sufficient")
-      .replace(
-        "Overall tier: F\nProof tier: F\nPatch tier: F",
-        "Overall tier: D\nProof tier: D\nPatch tier: D",
-      );
-    const synced = reportWithSyncedReviewComment(sourceReport, 343, "none");
-    writeFileSync(join(itemsDir, "343.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 343,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Canonical PR still computing mergeability",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: null,
-            labels: ["proof: sufficient"],
-          },
-        },
-      }),
-      () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: [
-            "--target-repo",
-            "openclaw/openclaw",
-            "--dry-run",
-            "--apply-kind",
-            "all",
-            "--processed-limit",
-            "3",
-          ],
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions does not promote PRs superseded by non-clean linked pull requests", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const sourceReport = stalePullRequestReport({
-      number: 345,
-      title: "Old activity PR",
-      labels: JSON.stringify([]),
-      pr_rating_overall: "D",
-      pr_rating_proof: "D",
-      pr_rating_patch: "D",
-      work_cluster_refs: JSON.stringify([
-        "Superseded by https://github.com/openclaw/openclaw/pull/400",
-      ]),
-    })
-      .replace("Status: missing", "Status: sufficient")
-      .replace(
-        "Overall tier: F\nProof tier: F\nPatch tier: F",
-        "Overall tier: D\nProof tier: D\nPatch tier: D",
-      );
-    const synced = reportWithSyncedReviewComment(sourceReport, 345, "none");
-    writeFileSync(join(itemsDir, "345.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 345,
-        title: "Old activity PR",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Blocked canonical PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "open",
-            merged_at: null,
-            mergeable_state: "blocked",
-            labels: ["proof: sufficient"],
-          },
-        },
-      }),
-      () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: [
-            "--target-repo",
-            "openclaw/openclaw",
-            "--dry-run",
-            "--apply-kind",
-            "all",
-            "--processed-limit",
-            "3",
-          ],
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{ action: string }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions blocks duplicate close when linked canonical PR closed unmerged", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 336,
-        title: "Already proposed duplicate close",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }),
-      336,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "336.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 336,
-        title: "Already proposed duplicate close",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Closed unmerged canonical PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: null,
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: [
-            "--target-repo",
-            "openclaw/openclaw",
-            "--dry-run",
-            "--apply-kind",
-            "all",
-            "--processed-limit",
-            "3",
-          ],
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "kept_open")?.reason ?? "",
-      /closed and unmerged/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions blocks duplicate close when canonical PR is only in close comment", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const reportMarkdown = lowSignalCloseReport({
-      number: 346,
-      title: "Already proposed duplicate close",
-      close_reason: "duplicate_or_superseded",
-      work_cluster_refs: JSON.stringify([]),
-    }).replace(
-      "Closing this PR because the branch is not a useful landing base.",
-      [
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-        "",
-        "Earlier context also mentioned https://github.com/openclaw/openclaw/pull/401.",
-      ].join("\n"),
-    );
-    const synced = reportWithSyncedReviewComment(reportMarkdown, 346, "duplicate_or_superseded");
-    writeFileSync(join(itemsDir, "346.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 346,
-        title: "Already proposed duplicate close",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Closed unmerged canonical PR",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: null,
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        runApplyDecisionsForTest({
-          itemsDir,
-          closedDir,
-          plansDir,
-          reportPath,
-          extraArgs: [
-            "--target-repo",
-            "openclaw/openclaw",
-            "--dry-run",
-            "--apply-kind",
-            "all",
-            "--processed-limit",
-            "3",
-          ],
-        });
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.match(
-      report.find((entry) => entry.action === "kept_open")?.reason ?? "",
-      /closed and unmerged/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions keeps existing duplicate PR close proposals open when coverage proof says keep_open", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const commentWriteLogPath = join(root, "comment-write.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 348,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Superseded by https://github.com/openclaw/openclaw/pull/400",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-      ),
-      348,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "348.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 348,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        commentWriteLogPath,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Cleans up provider setup without changing the fallback route.",
-            comments: [
-              {
-                id: 9400,
-                html_url: "https://github.com/openclaw/openclaw/pull/400#issuecomment-9400",
-                created_at: "2026-05-01T02:00:00Z",
-                updated_at: "2026-05-01T02:00:00Z",
-                user: { login: "maintainer" },
-                body: "This does not include the fallback route behavior from PR 348.",
-              },
-            ],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "keep_open",
-            reason: "PR A still has unique fallback route behavior that PR B does not cover.",
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.equal(
-      report.find((entry) => entry.number === 348)?.action,
-      "skipped_pr_close_coverage_proof",
-    );
-    assert.match(
-      report.find((entry) => entry.action === "skipped_pr_close_coverage_proof")?.reason ?? "",
-      /unique fallback route behavior/,
-    );
-    assert.match(
-      readFileSync(join(itemsDir, "348.md"), "utf8"),
-      /action_taken: skipped_pr_close_coverage_proof/,
-    );
-    const blockedReport = readFileSync(join(itemsDir, "348.md"), "utf8");
-    assert.match(blockedReport, /^decision: keep_open$/m);
-    assert.match(blockedReport, /^close_reason: none$/m);
-    assert.match(blockedReport, /## PR Close Coverage Proof\n\nDecision: keep_open/);
-    assert.match(blockedReport, /unique fallback route behavior/);
-    assert.match(readFileSync(commentWriteLogPath, "utf8"), /issues\/comments\/9348/);
-    assert.doesNotMatch(
-      renderReviewCommentFromReport(blockedReport, "none"),
-      /I’m closing this PR/,
-    );
-
-    writeFileSync(commentWriteLogPath, "", "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 348,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        commentWriteLogPath,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Cleans up provider setup without changing the fallback route.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(root, { type: "failure", message: "proof should not rerun" }, () => {
-          runApplyDecisionsForTest({
-            itemsDir,
-            closedDir,
-            plansDir,
-            reportPath,
-            extraArgs: [
-              "--target-repo",
-              "openclaw/openclaw",
-              "--apply-kind",
-              "all",
-              "--sync-comments-only",
-              "--processed-limit",
-              "3",
-            ],
-          });
-        });
-      },
-    );
-
-    const retryReport = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      number: number;
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      retryReport.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.equal(
-      retryReport.find((entry) => entry.number === 348)?.action,
-      "review_comment_synced",
-    );
-    assert.equal(
-      retryReport.some((entry) => /proof should not rerun/.test(entry.reason)),
-      false,
-    );
-    assert.match(readFileSync(commentWriteLogPath, "utf8"), /issues\/comments\/9348/);
-    assert.equal(existsSync(join(closedDir, "348.md")), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions skips duplicate PR coverage proof during synced comment-only runs", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const reportMarkdown = lowSignalCloseReport({
-      number: 348,
-      title: "Provider route fallback",
-      close_reason: "duplicate_or_superseded",
-      work_cluster_refs: JSON.stringify([
-        "Superseded by https://github.com/openclaw/openclaw/pull/400",
-      ]),
-    }).replace(
-      "Closing this PR because the branch is not a useful landing base.",
-      "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-    );
-    const synced = reportWithSyncedReviewComment(reportMarkdown, 348, "duplicate_or_superseded");
-    writeFileSync(join(itemsDir, "348.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 348,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Cleans up provider setup without changing the fallback route.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          { type: "failure", message: "proof should not run", invocationLogPath: proofLogPath },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--sync-comments-only",
-                "--processed-limit",
-                "1",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    assert.equal(existsSync(proofLogPath), false);
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => /proof should not run/.test(entry.reason)),
-      false,
-    );
-    assert.equal(existsSync(join(closedDir, "348.md")), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions skips duplicate PR coverage proof during stale comment-only sync", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    const commentWriteLogPath = join(root, "comment-write.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const reportMarkdown = lowSignalCloseReport({
-      number: 349,
-      title: "Provider route fallback",
-      close_reason: "duplicate_or_superseded",
-      work_cluster_refs: JSON.stringify([
-        "Superseded by https://github.com/openclaw/openclaw/pull/400",
-      ]),
-    }).replace(
-      "Closing this PR because the branch is not a useful landing base.",
-      "Closing this PR as superseded by https://github.com/openclaw/openclaw/pull/400.",
-    );
-    const synced = reportWithSyncedReviewComment(reportMarkdown, 349, "duplicate_or_superseded");
-    writeFileSync(join(itemsDir, "349.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 349,
-        title: "Provider route fallback",
-        comment: markedReviewCommentForTest(349, "Stale durable review comment."),
-        commentWriteLogPath,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Cleans up provider setup without changing the fallback route.",
-            comments: [],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          { type: "failure", message: "proof should not run", invocationLogPath: proofLogPath },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--sync-comments-only",
-                "--processed-limit",
-                "1",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    assert.equal(existsSync(proofLogPath), false);
-    assert.match(readFileSync(commentWriteLogPath, "utf8"), /issues\/comments\/9349/);
-    assert.deepEqual(JSON.parse(readFileSync(reportPath, "utf8")), [
-      {
-        number: 349,
-        action: "review_comment_synced",
-        reason: "updated durable Codex review comment",
-      },
-    ]);
-    assert.equal(existsSync(join(closedDir, "349.md")), false);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions gates duplicate PR closes with shorthand canonical refs", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 356,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify(["Superseded by #400"]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR as superseded by openclaw/openclaw#400.",
-      ),
-      356,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "356.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 356,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Cleans up provider setup without changing the fallback route.",
-            comments: [
-              {
-                id: 9400,
-                html_url: "https://github.com/openclaw/openclaw/pull/400#issuecomment-9400",
-                created_at: "2026-05-01T02:00:00Z",
-                updated_at: "2026-05-01T02:00:00Z",
-                user: { login: "maintainer" },
-                body: "This does not include the fallback route behavior from PR 356.",
-              },
-            ],
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "keep_open",
-            reason: "PR A still has unique fallback route behavior that PR B does not cover.",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      number: number;
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.equal(
-      report.find((entry) => entry.number === 356)?.action,
-      "skipped_pr_close_coverage_proof",
-    );
-    assert.match(readFileSync(proofLogPath, "utf8"), /proof/);
-    assert.match(
-      report.find((entry) => entry.number === 356)?.reason ?? "",
-      /unique fallback route behavior/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("apply-decisions gates duplicate PR closes when unrelated bare issue refs accompany one PR URL", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const closedDir = join(root, "closed");
-    const plansDir = join(root, "plans");
-    const reportPath = join(root, "apply-report.json");
-    const proofLogPath = join(root, "proof.log");
-    mkdirSync(itemsDir, { recursive: true });
-    mkdirSync(plansDir, { recursive: true });
-    const synced = reportWithSyncedReviewComment(
-      lowSignalCloseReport({
-        number: 359,
-        title: "Provider route fallback",
-        close_reason: "duplicate_or_superseded",
-        work_cluster_refs: JSON.stringify([
-          "Related pull request: https://github.com/openclaw/openclaw/pull/400",
-          "Background issue: #500",
-        ]),
-      }).replace(
-        "Closing this PR because the branch is not a useful landing base.",
-        "Closing this PR because the related pull request is the better review target.",
-      ),
-      359,
-      "duplicate_or_superseded",
-    );
-    writeFileSync(join(itemsDir, "359.md"), synced.report, "utf8");
-
-    withMockGh(
-      root,
-      promotionGhMock({
-        number: 359,
-        title: "Provider route fallback",
-        comment: synced.comment,
-        linkedPulls: {
-          400: {
-            number: 400,
-            title: "Provider cleanup",
-            html_url: "https://github.com/openclaw/openclaw/pull/400",
-            state: "closed",
-            merged_at: "2026-05-02T00:00:00Z",
-            body: "Cleans up provider setup without changing the fallback route.",
-            comments: [],
-            labels: [],
-          },
-        },
-        linkedIssues: {
-          500: {
-            number: 500,
-            title: "Related provider issue",
-            html_url: "https://github.com/openclaw/openclaw/issues/500",
-            state: "open",
-            labels: [],
-          },
-        },
-      }),
-      () => {
-        withMockCodexProof(
-          root,
-          {
-            type: "decision",
-            decision: "keep_open",
-            reason: "PR A still has unique fallback route behavior that PR B does not cover.",
-            invocationLogPath: proofLogPath,
-          },
-          () => {
-            runApplyDecisionsForTest({
-              itemsDir,
-              closedDir,
-              plansDir,
-              reportPath,
-              extraArgs: [
-                "--target-repo",
-                "openclaw/openclaw",
-                "--apply-kind",
-                "all",
-                "--processed-limit",
-                "3",
-              ],
-            });
-          },
-        );
-      },
-    );
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      number: number;
-      action: string;
-      reason: string;
-    }>;
-    assert.equal(
-      report.some((entry) => entry.action === "closed"),
-      false,
-    );
-    assert.equal(
-      report.find((entry) => entry.number === 359)?.action,
-      "skipped_pr_close_coverage_proof",
-    );
-    assert.match(readFileSync(proofLogPath, "utf8"), /proof/);
-    assert.match(
-      report.find((entry) => entry.number === 359)?.reason ?? "",
-      /unique fallback route behavior/,
-    );
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
 
 test("apply-decisions ignores bare refs inside cross-repo markdown link labels for duplicate proof", () => {
   const root = mkdtempSync(tmpPrefix);
